requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
PyMuPDF
google-generativeai
python-multipart